Changes:
- Qty button:
In this case I created a new View: QtyButton.
With this view, I defined states for quantity (zero-one-normal-max). With these states I can set different colors or drawables using selectors.
On the other hand, this button also receives as a parameter in xml a btn_type in order to set as add or reduce and disable button based on needings.

- Label:
Another attr created: colorsRandom.
ColorsRandom is an array of colors, and is defined on every theme to set the colors that can take the label.
I created a defaults_theme.xml so that light and dark themes can extend from it and have default values for the custom attributes defined. Otherwise if I forgot to set attr on theme, app would crash.
Finally, colorsRandom array and other defaults are gotten from resources from theme in Utils ui prods.

- TODOs mainProdsItem:
Load image using Coil or Glide.
Spans on product description.

------------------------------------------------------

- Additional small changes:
ConstantsProducts was renamed and organized as Constants.Products in core package.
Inserted a new graphik.xml font-family with weights, but could be not necessary.
ic_cart/ic_person/ic_store state_selected false removed. The item with no states in selector is considered the default.