Changes:
1. Added recyclerView to show the variations: Additionally I used different viewHolders.
2. Changed name ProdWithQuantity to ProdWithQuantities, as quantity is a map that can have more than one key.
3. Created drawables ic_add_variation and ic_reduce_variation to define background of buttons quantities variations.
4. Changes on quantity button: Now inherits from MaterialButton and not from AppCompatButton, and I did some refactoring.
5. Changed adapter initialization of recyclerViews (mainProducts and variations) to lazy.
6. Change in theme: shapeLargeComponent. With percentage value, borders were extremely rounded on high dialog, so I used dp value.

------------------------------------------------

DETAILS:
1. The recyclerView is able to show a list with headers in case any of the variations has details, or a basic list with no details, when none of the variation has details.
I did that by defining three viewHolders: Header and ItemForHeader represent the elements for list of variations with details, and BasicItem when none of the variations has details.
VariationUi is like a supporting class, all the information from the variation in the viewModel is placed in those components that help organize where to put information in header.

4. QtyButton Changes:
I made it extend from MaterialButton as that would be more coherent with the theme, and could give some added values.
But the main reason was that when setting enabled or disabled to the button, if it was extending from AppCompatButton I have to also manage changing the alpha and the color, with MaterialButton instead, this behaviour is already implemented and looks according to the theme.
Also created method to setQuantityState based on quantity value and maxOrdervalue, and made some refactoring to setNewEnabledState based on quantityState, cleaning the setQuantityState method and making more ndestandable the method setNewEnabledState.

5. I think creation of the adapter was being done with the creation of the fragment. If I made the instantiation outside the onCreateView method there could be memory leaks, but using by lazy, it would be created when needed so it is also ok.

------------------------------------------------

coming TODOs:
- Design: Spaces, colors (also ColorStateList for buttonsQuantities), visibility of elements in list_item_variation.
- Loading and failure states.
- Functionality to buttons to add or reduce products.