Navigation for user login flow working.
Features:
- Navigating with bottomNav:
  -> If user is logged in, show user data (do nothing).
  -> If not logged in, before creating the view navigate to login flow.

- Pressing back / cancelling login flow will not show user data.
  -> It will also pop this fragment from back stack, just if cancelling login flow.

- Completing login flow successfully will take user to the fragment that call the login flow (user/cart/...).
  -> User will return to the fragment that call the login flow to start. Normally the userData fragment, but could be cart fragment when verifying user is signedIn.

------------------------------------
How? Changes:
The idea is to get a result in the fragment that call the login flow: false if cancelled, true if completed successfully.

UserFragment: Observe the result by observing its stateHandle.
SignInOptionsFragment: Get the stateHandle from prevBackStackEntry, and set false when flow is started.
SignInFragments: Can set true on stateHandle of the UserFragment.kt
To get the stateHandle of the userFragment, I saved the destination id in the stateHandle of a known destination: the start destination of the login flow.
With that destination, I can get the stateHandle and modify to send the result back to the fragment that started login flow.

There were also some other small improvements:
- Names to functions parameters in ErrorHelpers.kt
- Removing condition visibility gone on userData fragment, as navigation was moved to onCreate (before view is actually created)
- Fixed nav_main.xml:
  -> Removing popUpTo condition in action navigate userData to loginFlow as it is now handled differently (before when login was successful didn't show userData).
- Removed signIn/signUp destinations from nav_sign_in.xml as they weren't being used (they are contained in emailFragment destination.