Small changes: Basically I renamed some classes and files to create an auth package/module.
For example I decided to call user in auth Module as Profile.kt.

As a general rule, I try to link packages just in the domain/use cases. I must not call a class from another package in data nor ui layer.

- Implemented FirebaseCloudMessaging to display notifications.
At first time it was not working, but with the last tests, notifications were actually being shown.