UserData is now also saved in firestore database and edit user data is now working correctly.
This, in order to later saved additional user information, such as phoneNumber and address.
There could still be room for some improvement, like avoid calling update in firestore on every signIn/signUp with facebook or google, but the problem would be that access token in facebook would not be renovated.

If that is not a problem, I could use the additionalUserInfo in repo, from authResult, in order to know if user is new, and then send to useCase which will decide to updateFirestore or not.