Bug corrected: Delete address is now working properly.
The problem was because DtosTransformations was generating a new Id every time I called to transform Address into AddressDto, so the id in the AddressObject got lost, and the address to delete was therefore not the same. I just changed the to get the same id coming from Address (important as object was already taken from firestore in delete), and the generation of the new Id is placed in the use case.