Improved ui on makeOrderButton in cart. Button is now visually different when is not enabled.
Also improved code when getting colorFromAttr in observers. In CartFragment.kt and ShippingInfoFragment.kt I created classes that help calling all the logic in the observer, and save the state in the class internally, so if the state has not changed, it doesn't bother with calling the ui to change.