Corrected fetching from products using local cache room.
-> Products were always being fetch from remote, as I was taking the lastUpdatedTime from the actualLastUpdatedTime, in other words the last time admin change a product.
It was improved by taking the last successful fetch the device made to the remote database, even when this last fetch didn't updated nothing because all was actually up to date, the only condition is that this last fetch was successful.

It has also improved the speed on getting variations and in addition, it won't request remote database every time a product is clicked in order to get the variations.

With this change, search was also made a lot faster and cheaper.